Add new feature to process user input
Integrated blockchain functionality for the credentials storing. Both database and blockchain data persistence are available. Small caveat that the bigger the blockchain gets, the harder for the data integrity verification it will get. As a PoC it works but will need further development and testing for a future production release